fix(macro): parse scoped historical FOMC votes
